update echo example

+++ src/include/packet_context.h
@@ -1,7 +1,6 @@
 #ifndef PACKET_CONTEXT_H
 #define PACKET_CONTEXT_H
 
-
 // Return values
 #define PASS 100
 #define DROP 200
@@ -11,7 +10,7 @@
 struct pktctx {
 	void *data;
 	void *data_end;
-	size_t pkt_len;
+	uint64_t pkt_len;
 };
 
 #endif
